separated row component from board component. However, importing row component requires specifying the path './Row' as opposed to simply 'Row'. Importing style sheet also requires using the css modules method where style is import like so 'import style from css path'. Unsure why on both points
