+++ build-dev/client/index.html
@@ -8,5 +8,5 @@
 <body>
 <h1>Wow, react</h1>
 <div id="app"></div>
-</body>
+<script type="text/javascript" src="/main.js"></script></body>
 </html>
