+++ src/client/App.jsx
@@ -2,11 +2,12 @@ import React from 'react';
 import { hot } from 'react-hot-loader';
 
 import Board from './components/board/board';
+import styles from './style.scss';
 
 class App extends React.Component {
   render() {
     return (
-      <div>
+      <div className={styles.hello}>
         <p>Welcome.</p>
         <Board/>
       </div>
