+++ src/client/components/board/board.scss
@@ -1,0 +1,10 @@
+.boo{
+	border: solid black;
+	display: inline-block;
+	padding: 2px;
+	font-size: 30px;
+}
+
+.row{
+	margin: 0px;
+}
