+++ src/client/style.scss
@@ -1,13 +1,11 @@
+
 html, body {
   background-color: white;
   min-width: 700px;
   font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
 
 }
+
 body{
   padding:20px;
-
-}
-p{
-	color:red;
 }
